Allow compilation in ILP64 mode
The code now uses preprocessor macro `Int` instead of the `int` type
to allow for flexibility in compilation of the library. `Int` is the
internal integer type used for most operations. Also, it is the type
used (expected) by the library interface. By default `Int` is `int`.
However, it is possible to redefine it on the command line to a
larger integer type `long` if desired.

Any change to `Int` needs to be accompanied with the appropriate
promotion of the Fortran integers (`-fdefault-integer-8`, `-i8`
etc.).

When compiling ILP64 version of ScaLAPACK, the following needs to be
satisfied:
  1. Use ILP64 BLAS (e.g. OpenBLAS with `-DINTERFACE64=ON`).
  2. Promote integer types (`-DInt=long`, `-fdefault-integer-8`).
  3. Change hardcoded `INTSZ` and `INTGSZ` in the test sources
     from 4 to 8.

+++ REDIST/SRC/pdgemr.c
@@ -158,20 +158,20 @@
 #define Clacpy Cdgelacpy
 void  Clacpy();
 typedef struct {
-  int   desctype;
-  int   ctxt;
-  int   m;
-  int   n;
-  int   nbrow;
-  int   nbcol;
-  int   sprow;
-  int   spcol;
-  int   lda;
+  Int   desctype;
+  Int   ctxt;
+  Int   m;
+  Int   n;
+  Int   nbrow;
+  Int   nbcol;
+  Int   sprow;
+  Int   spcol;
+  Int   lda;
 }     MDESC;
 #define BLOCK_CYCLIC_2D 1
 typedef struct {
-  int   lstart;
-  int   len;
+  Int   lstart;
+  Int   len;
 }     IDESC;
 #define SHIFT(row,sprow,nbrow) ((row)-(sprow)+ ((row) >= (sprow) ? 0 : (nbrow)))
 #define max(A,B) ((A)>(B)?(A):(B))
@@ -185,7 +185,7 @@ typedef struct {
 #endif
 /* Cblacs */
 extern void Cblacs_pcoord();
-extern int Cblacs_pnum();
+extern Int Cblacs_pnum();
 extern void Csetpvmtids();
 extern void Cblacs_get();
 extern void Cblacs_pinfo();
@@ -206,12 +206,12 @@ extern void Cdgerv2d();
 /* lapack */
 void  dlacpy_();
 /* aux fonctions */
-extern int localindice();
+extern Int localindice();
 extern void *mr2d_malloc();
-extern int ppcm();
-extern int localsize();
-extern int memoryblocksize();
-extern int changeorigin();
+extern Int ppcm();
+extern Int localsize();
+extern Int memoryblocksize();
+extern Int changeorigin();
 extern void paramcheck();
 /* tools and others function */
 #define scanD0 dgescanD0
@@ -223,7 +223,7 @@ extern void scanD0();
 extern void dispmat();
 extern void setmemory();
 extern void freememory();
-extern int scan_intervals();
+extern Int scan_intervals();
 extern void Cpdgemr2do();
 extern void Cpdgemr2d();
 /* some defines for Cpdgemr2do */
@@ -243,8 +243,8 @@ extern void Cpdgemr2d();
 void 
 fortran_mr2d(m, n, A, ia, ja, desc_A,
 	     B, ib, jb, desc_B)
-  int  *ia, *ib, *ja, *jb, *m, *n;
-  int   desc_A[DESCLEN], desc_B[DESCLEN];
+  Int  *ia, *ib, *ja, *jb, *m, *n;
+  Int   desc_A[DESCLEN], desc_B[DESCLEN];
   double *A, *B;
 {
   Cpdgemr2do(*m, *n, A, *ia, *ja, (MDESC *) desc_A,
@@ -254,18 +254,18 @@ fortran_mr2d(m, n, A, ia, ja, desc_A,
 void 
 fortran_mr2dnew(m, n, A, ia, ja, desc_A,
 		B, ib, jb, desc_B, gcontext)
-  int  *ia, *ib, *ja, *jb, *m, *n;
-  int   desc_A[DESCLEN], desc_B[DESCLEN];
+  Int  *ia, *ib, *ja, *jb, *m, *n;
+  Int   desc_A[DESCLEN], desc_B[DESCLEN];
   double *A, *B;
-  int  *gcontext;
+  Int  *gcontext;
 {
   Cpdgemr2d(*m, *n, A, *ia, *ja, (MDESC *) desc_A,
 	    B, *ib, *jb, (MDESC *) desc_B, *gcontext);
   return;
 }
 static2 void init_chenille();
-static2 int inter_len();
-static2 int block2buff();
+static2 Int inter_len();
+static2 Int block2buff();
 static2 void buff2block();
 static2 void gridreshape();
 void
@@ -276,15 +276,15 @@ Cpdgemr2do(m, n,
 /* pointers to the memory location of the matrix and the redistributed matrix */
   MDESC *ma;
   MDESC *mb;
-  int   ia, ja, ib, jb, m, n;
+  Int   ia, ja, ib, jb, m, n;
 {
-  int   dummy, nprocs;
-  int   gcontext;
+  Int   dummy, nprocs;
+  Int   gcontext;
   /* first we initialize a global grid which serve as a reference to
    * communicate from grid a to grid b */
   Cblacs_pinfo(&dummy, &nprocs);
-  Cblacs_get(0, 0, &gcontext);
-  Cblacs_gridinit(&gcontext, "R", 1, nprocs);
+  Cblacs_get((Int)0, (Int)0, &gcontext);
+  Cblacs_gridinit(&gcontext, "R", (Int)1, nprocs);
   Cpdgemr2d(m, n, ptrmyblock, ia, ja, ma,
 	    ptrmynewblock, ib, jb, mb, gcontext);
   Cblacs_gridexit(gcontext);
@@ -300,22 +300,22 @@ Cpdgemr2d(m, n,
 /* pointers to the memory location of the matrix and the redistributed matrix */
   MDESC *ma;
   MDESC *mb;
-  int   ia, ja, ib, jb, m, n, globcontext;
+  Int   ia, ja, ib, jb, m, n, globcontext;
 {
   double *ptrsendbuff, *ptrrecvbuff, *ptrNULL = 0;
   double *recvptr;
   MDESC newa, newb;
-  int  *proc0, *proc1, *param;
-  int   mypnum, myprow0, mypcol0, myprow1, mypcol1, nprocs;
-  int   i, j;
-  int   nprow, npcol, gcontext;
-  int   recvsize, sendsize;
+  Int  *proc0, *proc1, *param;
+  Int   mypnum, myprow0, mypcol0, myprow1, mypcol1, nprocs;
+  Int   i, j;
+  Int   nprow, npcol, gcontext;
+  Int   recvsize, sendsize;
   IDESC *h_inter;	/* to store the horizontal intersections */
   IDESC *v_inter;	/* to store the vertical intersections */
-  int   hinter_nb, vinter_nb;	/* number of intrsections in both directions */
-  int   dummy;
-  int   p0, q0, p1, q1;
-  int  *ra, *ca;
+  Int   hinter_nb, vinter_nb;	/* number of intrsections in both directions */
+  Int   dummy;
+  Int   p0, q0, p1, q1;
+  Int  *ra, *ca;
   /* end of variables */
   /* To simplify further calcul we change the matrix indexation from
    * 1..m,1..n (fortran) to 0..m-1,0..n-1 */
@@ -346,7 +346,7 @@ Cpdgemr2d(m, n,
   assert((myprow1 < p1 && mypcol1 < q1) || (myprow1 == -1 && mypcol1 == -1));
   /* exchange the missing parameters among the processors: shape of grids and
    * location of the processors */
-  param = (int *) mr2d_malloc(3 * (nprocs * 2 + NBPARAM) * sizeof(int));
+  param = (Int *) mr2d_malloc(3 * (nprocs * 2 + NBPARAM) * sizeof(Int));
   ra = param + nprocs * 2 + NBPARAM;
   ca = param + (nprocs * 2 + NBPARAM) * 2;
   for (i = 0; i < nprocs * 2 + NBPARAM; i++)
@@ -381,8 +381,8 @@ Cpdgemr2d(m, n,
     param[18] = ib;
     param[19] = jb;
   }
-  Cigamn2d(gcontext, "All", "H", 2 * nprocs + NBPARAM, 1, param, 2 * nprocs + NBPARAM,
-	   ra, ca, 2 * nprocs + NBPARAM, -1, -1);
+  Cigamn2d(gcontext, "All", "H", 2 * nprocs + NBPARAM, (Int)1, param, 2 * nprocs + NBPARAM,
+	   ra, ca, 2 * nprocs + NBPARAM, (Int)-1, (Int)-1);
   newa = *ma;
   newb = *mb;
   ma = &newa;
@@ -428,7 +428,7 @@ Cpdgemr2d(m, n,
   paramcheck(mb, ib, jb, m, n, p1, q1, gcontext);
   /* we change the problem so that ia < a->nbrow ... andia + m = a->m ... */
   {
-    int   decal;
+    Int   decal;
     ia = changeorigin(myprow0, ma->sprow, p0,
 		      ma->nbrow, ia, &decal, &ma->sprow);
     ptrmyblock += decal;
@@ -482,9 +482,9 @@ Cpdgemr2d(m, n,
    * of recvbuff the right place (scanD)(RECVBUFF)) */
   recvptr = ptrrecvbuff;
   {
-    int   tot, myrang, step, sens;
-    int  *sender, *recver;
-    int   mesending, merecving;
+    Int   tot, myrang, step, sens;
+    Int  *sender, *recver;
+    Int   mesending, merecving;
     tot = max(p0 * q0, p1 * q1);
     init_chenille(mypnum, nprocs, p0 * q0, proc0, p1 * q1, proc1,
 		  &sender, &recver, &myrang);
@@ -519,8 +519,8 @@ Cpdgemr2d(m, n,
 	  if (sendsize > 0
 	      && (step != myrang || !merecving)
 		) {
-	    Cdgesd2d(gcontext, sendsize, 1, ptrsendbuff, sendsize,
-		     0, proc1[i * q1 + j]);
+	    Cdgesd2d(gcontext, sendsize, (Int)1, ptrsendbuff, sendsize,
+		     (Int)0, proc1[i * q1 + j]);
 	  }	/* sendsize > 0 */
 	}	/* if (mesending ... */
 	if (merecving && sender[step] >= 0 &&
@@ -538,8 +538,8 @@ Cpdgemr2d(m, n,
 		     ptrsendbuff, recvsize,
 		     ptrrecvbuff, recvsize);
 	    } else {
-	      Cdgerv2d(gcontext, recvsize, 1, ptrrecvbuff, recvsize,
-		       0, proc0[i * q0 + j]);
+	      Cdgerv2d(gcontext, recvsize, (Int)1, ptrrecvbuff, recvsize,
+		       (Int)0, proc0[i * q0 + j]);
 	    }
 	  }	/* recvsize > 0 */
 	}	/* if (merecving ...) */
@@ -570,13 +570,13 @@ after_comm:
 }/* distrib */
 static2 void 
 init_chenille(mypnum, nprocs, n0, proc0, n1, proc1, psend, precv, myrang)
-  int   nprocs, mypnum, n0, n1;
-  int  *proc0, *proc1, **psend, **precv, *myrang;
+  Int   nprocs, mypnum, n0, n1;
+  Int  *proc0, *proc1, **psend, **precv, *myrang;
 {
-  int   ns, nr, i, tot;
-  int  *sender, *recver, *g0, *g1;
+  Int   ns, nr, i, tot;
+  Int  *sender, *recver, *g0, *g1;
   tot = max(n0, n1);
-  sender = (int *) mr2d_malloc((nprocs + tot) * sizeof(int) * 2);
+  sender = (Int *) mr2d_malloc((nprocs + tot) * sizeof(Int) * 2);
   recver = sender + tot;
   *psend = sender;
   *precv = recver;
@@ -627,8 +627,8 @@ init_chenille(mypnum, nprocs, n0, proc0, n1, proc1, psend, precv, myrang)
 #define Mlacpy(mo,no,ao,ldao,bo,ldbo) \
 { \
 double *_a,*_b; \
-int _m,_n,_lda,_ldb; \
-    int _i,_j; \
+Int _m,_n,_lda,_ldb; \
+    Int _i,_j; \
     _m = (mo);_n = (no); \
     _a = (ao);_b = (bo); \
     _lda = (ldao) - _m; \
@@ -641,14 +641,14 @@ int _m,_n,_lda,_ldb; \
       _a += _lda; \
     } \
 }
-static2 int 
+static2 Int 
 block2buff(vi, vinb, hi, hinb, ptra, ma, buff)
-  int   hinb, vinb;
+  Int   hinb, vinb;
   IDESC *hi, *vi;
   MDESC *ma;
   double *buff, *ptra;
 {
-  int   h, v, sizebuff;
+  Int   h, v, sizebuff;
   double *ptr2;
   sizebuff = 0;
   for (h = 0; h < hinb; h++) {
@@ -665,12 +665,12 @@ block2buff(vi, vinb, hi, hinb, ptra, ma, buff)
 }
 static2 void 
 buff2block(vi, vinb, hi, hinb, buff, ptrb, mb)
-  int   hinb, vinb;
+  Int   hinb, vinb;
   IDESC *hi, *vi;
   MDESC *mb;
   double *buff, *ptrb;
 {
-  int   h, v, sizebuff;
+  Int   h, v, sizebuff;
   double *ptr2;
   sizebuff = 0;
   for (h = 0; h < hinb; h++) {
@@ -684,12 +684,12 @@ buff2block(vi, vinb, hi, hinb, buff, ptrb, mb)
     }
   }
 }
-static2 int 
+static2 Int 
 inter_len(hinb, hi, vinb, vi)
-  int   hinb, vinb;
+  Int   hinb, vinb;
   IDESC *hi, *vi;
 {
-  int   hlen, vlen, h, v;
+  Int   hlen, vlen, h, v;
   hlen = 0;
   for (h = 0; h < hinb; h++)
     hlen += hi[h].len;
@@ -701,9 +701,9 @@ inter_len(hinb, hi, vinb, vi)
 void 
 Clacpy(m, n, a, lda, b, ldb)
   double *a, *b;
-  int   m, n, lda, ldb;
+  Int   m, n, lda, ldb;
 {
-  int   i, j;
+  Int   i, j;
   lda -= m;
   ldb -= m;
   assert(lda >= 0 && ldb >= 0);
@@ -716,23 +716,23 @@ Clacpy(m, n, a, lda, b, ldb)
 }
 static2 void 
 gridreshape(ctxtp)
-  int  *ctxtp;
+  Int  *ctxtp;
 {
-  int   ori, final;	/* original context, and new context created, with
+  Int   ori, final;	/* original context, and new context created, with
 			 * line form */
-  int   nprow, npcol, myrow, mycol;
-  int  *usermap;
-  int   i, j;
+  Int   nprow, npcol, myrow, mycol;
+  Int  *usermap;
+  Int   i, j;
   ori = *ctxtp;
   Cblacs_gridinfo(ori, &nprow, &npcol, &myrow, &mycol);
-  usermap = mr2d_malloc(sizeof(int) * nprow * npcol);
+  usermap = mr2d_malloc(sizeof(Int) * nprow * npcol);
   for (i = 0; i < nprow; i++)
     for (j = 0; j < npcol; j++) {
       usermap[i + j * nprow] = Cblacs_pnum(ori, i, j);
     }
   /* Cblacs_get(0, 0, &final); */
-  Cblacs_get(ori, 10, &final);
-  Cblacs_gridmap(&final, usermap, 1, 1, nprow * npcol);
+  Cblacs_get(ori, (Int)10, &final);
+  Cblacs_gridmap(&final, usermap, (Int)1, (Int)1, nprow * npcol);
   *ctxtp = final;
   free(usermap);
 }
